+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added a debug storage disk
 - Added tooltip to solderer progress bar that shows progress percentage
 - Added support for the Tesla energy system
+- Added support for the IC2 (EU) energy system
 - Tweaked grid GUI
 
 ### 0.8.3